Reservation :  - Ajout de la présence de la signature pour confirmer la reservation

+++ js/Reservation.js
@@ -4,13 +4,10 @@ class Reservation {
         this.input_firstname = $('#firstname');//Champ "prénom" de résa
 
         this.station_reservation = $('#name_station');
-        this.input_reservation_form = $('#btn_reservation_form');//Bouton de résa
 
         this.session_firstname = localStorage.getItem("prenom");
         this.session_lastname = localStorage.getItem("nom");
         this.session_timer = localStorage.getItem("endTimerReservation");
-
-        //localStorage.clear();
 
         this.initSettings();
 
@@ -18,13 +15,10 @@ class Reservation {
         document.getElementById('btn_reservation_form').addEventListener('click', function () {
             reservation.reservationCheck();
         });
-
-
     } // Fin du constructeur
 
     initSettings() {
             if (!this.storageAvailable('localStorage')) {
-
                 console.log("Impossible d'utiliser local storage!");
             }
             if (!localStorage.nom) { // s'il n'y a pas d'élément name, on laisse l'utilisateur faire
@@ -48,16 +42,23 @@ class Reservation {
 
     reservationCheck() {
         if (reservation.input_lastname.val() !== "" || reservation.input_firstname.val() !== "") {
-            // Nettoyage des valeurs du local et session storage pour un nouvel enregistrement
-             localStorage.clear();
-             sessionStorage.clear();
-            // Enregistrement des nouvelles valeurs
-            localStorage.setItem("prenom", reservation.input_firstname.val());
-            localStorage.setItem("nom", reservation.input_lastname.val());
-            sessionStorage.setItem("station", reservation.station_reservation.text()); // enregistre (temporairement) la valeur station
+            if (signe.valid === true){
+                // Nettoyage des valeurs du local et session storage pour un nouvel enregistrement
+                localStorage.clear();
+                sessionStorage.clear();
 
-            // //Démarrage du timer
-            reservation.start_timer();
+                // Enregistrement des nouvelles valeurs
+                localStorage.setItem("prenom", reservation.input_firstname.val());
+                localStorage.setItem("nom", reservation.input_lastname.val());
+                sessionStorage.setItem("station", reservation.station_reservation.text()); // enregistre (temporairement) la valeur station
+
+                // //Démarrage du timer
+                reservation.start_timer();
+            } else{
+                $('#confirmation').css("display", "block");
+                $('#confirmation').html("Signature invalide, veuillez recommencer");
+                $('#confirmation').css("color", "red");
+            }
         } else {
             $("#form-error-renseignement").css("display", "block");
         }
@@ -66,7 +67,6 @@ class Reservation {
     reservation_exist() {
 
         if (localStorage.length !== 0) {
-            //localStorage.clear();
             $("#lastnameSession").html(this.session_lastname);
             $("#firstnameSession").html(this.session_firstname);
             $("#lastname").val(this.session_lastname);
@@ -74,11 +74,6 @@ class Reservation {
             $("#minutely").html(this.session_timer);
             localStorage.getItem("endTimerReservation");
             this.continue_timer();
-
-            //this.start_timer();
-            //console.log(localStorage.getItem("finishTime"));
-            //reservation.convert_milliseconds(localStorage.getItem("finishTime"));
-            //reservation.convert_milliseconds(localStorage.getItem("endTimerReservation"));
         }
     }
 
@@ -86,7 +81,6 @@ class Reservation {
         let date = new Date(time);
         let min = date.getMinutes();
         let sec = date.getSeconds();
-
 
         min = (min < 10) ? "0" + min : min;
         sec = (sec < 10) ? "0" + sec : sec;
@@ -98,8 +92,6 @@ class Reservation {
         // si le localstorage contient un temps (précedemment sauvegardé),
         if (localStorage.getItem("finishTime") !== null) {
             let add_time_at_reservation = new Date().getTime() + localStorage.getItem("finishTime");
-            console.log('date déja enregistré');
-
             // add_time_at_reservation sera à l'heure actuelle + le reste de temps du timer (donc du paramètre)
         }
 
@@ -110,6 +102,7 @@ class Reservation {
     start_timer(){
         // add_time_reservation sera à l'heure actuelle + 20 mins
         let add_time_at_reservation = new Date().getTime() + 1200000;
+
         //Stockage dans le localstorage du temps
         localStorage.setItem("finishTime", add_time_at_reservation.toString());
 
@@ -118,25 +111,29 @@ class Reservation {
     }
 
     decrement_time(){
+        $('#session').css("display", "block");
         $("#btn_reservation_form").css("display", "none");
         $("#btn_canceled_reservation").css("display", "block");
         let x = setInterval(function () {
             //Récuperation de l'heure actuelle
             let now = new Date().getTime();
-            //let convert= reservation.convert_milliseconds(now);
-            //console.log(localStorage);
+
             //Calcl du temps restant en soustrayant l'heure de reservation +20min à l'heure actuelle
             let time_stop = localStorage.getItem("finishTime") - now;
+
             //Enregistrement dans le localstorage
             localStorage.setItem("endTimerReservation", time_stop.toString());
+
             // Convertion du calcul en min et sec
             let returnTime = reservation.convert_milliseconds(time_stop);
+
             //Affichage dans la page du décompte
             $('#minutely').html( "Il vous reste " + returnTime + " avant la fin de la reservation");
             $('#lastnameSession').css("display", "block");
             $('#firstnameSession').css("display", "block");
             $('#lastnameSession').html("Réservation efféctué par " + localStorage.getItem("nom"));
             $('#firstnameSession').html(localStorage.getItem("prenom") + " à la station " + sessionStorage.getItem("station"));
+
             //Si le compte a rebours arrive a 00:00 alors stop la reservation
             if (returnTime === "00:00"){
                 clearInterval(x);
